Make native integration automatic rather than opt-in
